+++ R/global.R
@@ -2,5 +2,6 @@
 utils::globalVariables(
 	c(
 		"nc", "desc", "BIC_vals",
-		"model", "n", "v", "node"
+		"model", "n", "v", "node",
+		"P1", "P2"
 ))