+++ NEWS.md
@@ -1,8 +1,9 @@
-# mulgar 1.0.1.9000
+# mulgar 1.0.2
 
 * Fix with is_tibble
 * Added function to make a slice for a single projection
 * Added choice of number of PCs in ggscree
+* Additional data sets
 
 # mulgar 1.0.1